pyproject.toml
[tool.poetry]
name = "rag-unitesting"
version = "0.1"
description = "Creatig rag pipline using chaining and also test it using unitesting."
authors = ["Amrendra Singh"]
license = "Mit"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12"
numpy = "^2.4.0"
openai = "^2.14.0"
scikit-learn = "^1.8.0"
langchain = "^0.4.0"
langchain-core = "^1.2.0"
langchain-openai = "^1.1.0"
langchain-community = "^0.4.0"
langchain-text-splitters = "^1.1.0"
langsmith = "^0.6.0"
fastapi = "^0.128.0"
uvicorn = "^0.40.0"
pydantic = "^2.12.0"
python-dotenv = "^1.1.0"
pymilvus = "^2.6.0"
pypdf = "^4.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
